Release identifiers, video and company in postgres. Database Batch insert.
@philipmat 

Some issues:
- number of tracks in the DB is 53,434,173 while the number of XML
<track> elements is 53,832,994. Would be interesting to figure out where
these tracks go. Hopefully it is not because of my changes :)
- it would be nice to normalize the data (I write it more or less in the
way we get it from the XML) but this is not always possible; for
company, multiple companies have the ID of 0. Also, it seems that
different spellings apply for the same ID.
- batch import did not change so much the performance. It still took me
about 10 hours, but considering the added SQL. Probably dumping the data
in some file and perform bulk table loads is the long term solution.
- I run the code against Postgres 9.3, I removed some "set" statement
from the sql script, so that they run on my configuration
- I removed a try/finally because in case of an exception in the try
block, the finally got executed, but finally was throwing a new
exception. This way it is simpler to debug.